+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - richat: exit code 2 if replay is not available ([#177](https://github.com/lamports-dev/richat/pull/177))
 - filter: move fn update_write_version ([#178](https://github.com/lamports-dev/richat/pull/178))
 - filter: account keys limited parser ([#179](https://github.com/lamports-dev/richat/pull/179))
+- filter: add parse benchmark ([#180](https://github.com/lamports-dev/richat/pull/180))
 
 ### Breaking